refactor: start cleaning code, add docker-compose.yml
