+++ .env.sample
@@ -1,0 +1,24 @@
+# Domain
+DOMAIN=localhost
+
+# Environment: local / staging / production
+ENVIRONMENT=local
+
+# Backend
+BACKEND_CORS_ORIGINS="https://localhost"
+DOCKER_IMAGE_BACKEND=backend
+
+# Postgres
+POSTGRES_SERVER=localhost
+POSTGRES_PORT=5432
+POSTGRES_DB=promochator
+POSTGRES_USER=promochator
+POSTGRES_PASSWORD=changethis
+
+# Langchain
+LANGSMITH_API_KEY=changethis
+LANGSMITH_TRACING=true
+
+# LLMs
+OPENAI_API_KEY=changethis
+DOCKER_IMAGE_OLLAMA=ollama/ollama
