+++ scripts/ml/recomend.py
@@ -5,8 +5,8 @@ import fire
 import os
 from dotenv import load_dotenv
 
-from src.constants import PROMPT_TEMPLATE, VECTORSTORE_PATH
-from src.database import get_retriever
+from app.core.config import PROMPT_TEMPLATE, VECTORSTORE_PATH
+from app.core.database import get_retriever
 
 
 load_dotenv()
